[KLIB] Drop `KlibMetadataDeserializedPackageFragmentsFactory` class
- `createCachedPackageFragments()` is inlined to it's only use site.
- `createDeserializedPackageFragments()` is made a private top-level
  function and moved closer to it's single use site.

^KT-87387
