+++ compiler/util-klib-metadata/src/org/jetbrains/kotlin/library/metadata/KlibMetadataModuleDescriptorFactory.kt
@@ -20,7 +20,6 @@ import org.jetbrains.kotlin.storage.StorageManager
 interface KlibMetadataModuleDescriptorFactory {
 
     val descriptorFactory: KlibModuleDescriptorFactory
-    val packageFragmentsFactory: KlibMetadataDeserializedPackageFragmentsFactory
 
     @OptIn(K1Deprecation::class)
     val flexibleTypeDeserializer: FlexibleTypeDeserializer
